+++ client/bindings/nodejs/lib/Client.ts
@@ -20,7 +20,7 @@ import type {
     NftQueryParameter,
     AliasQueryParameter,
     LedgerNanoStatus,
-    IInputSigningData,
+    ISegment,
     OutputIdsResponse,
 } from '../types';
 import type {
@@ -384,18 +384,17 @@ export class Client {
      */
     async signatureUnlock(
         secretManager: SecretManager,
-        inputSigningData: IInputSigningData,
         // Uses `Array<number>` instead of `Uint8Array` because the latter serializes
         // as an object rather than an array, which results in errors with serde.
         transactionEssenceHash: Array<number>,
+        chain: ISegment[],
     ): Promise<UnlockTypes> {
         const response = await this.messageHandler.sendMessage({
             name: 'signatureUnlock',
             data: {
                 secretManager,
-                inputSigningData,
                 transactionEssenceHash,
-                remainderData: undefined,
+                chain
             },
         });