+++ client/src/secret/mod.rs
@@ -27,8 +27,8 @@ use iota_types::block::{
     address::Address,
     output::Output,
     payload::transaction::TransactionEssence,
-    signature::Ed25519Signature,
-    unlock::{AliasUnlock, NftUnlock, ReferenceUnlock, Unlock, Unlocks},
+    signature::{Ed25519Signature, Signature},
+    unlock::{AliasUnlock, NftUnlock, ReferenceUnlock, SignatureUnlock, Unlock, Unlocks},
 };
 use zeroize::ZeroizeOnDrop;
 
@@ -43,10 +43,9 @@ use crate::secret::types::StrongholdDto;
 use crate::{
     api::{
         input_selection::{is_alias_transition, Error as InputSelectionError},
-        PreparedTransactionData, RemainderData,
+        PreparedTransactionData,
     },
-    secret::types::InputSigningData,
-    unix_timestamp_now,
+    unix_timestamp_now, Error,
 };
 
 /// The secret manager interface.
@@ -64,16 +63,15 @@ pub trait SecretManage: Send + Sync {
         options: Option<GenerateAddressOptions>,
     ) -> crate::Result<Vec<Address>>;
 
-    /// Sign on `essence`, unlock `input` by returning an [Unlock].
-    async fn signature_unlock(
-        &self,
-        input: &InputSigningData,
-        essence_hash: &[u8; 32],
-        remainder: &Option<RemainderData>,
-    ) -> crate::Result<Unlock>;
-
-    /// Signs `msg` using the given `chain`.
+    /// Signs `msg` using the given [`Chain`].
     async fn sign_ed25519(&self, msg: &[u8], chain: &Chain) -> crate::Result<Ed25519Signature>;
+
+    /// Signs `essence_hash` using the given `chain`, returning an [`Unlock`].
+    async fn signature_unlock(&self, essence_hash: &[u8; 32], chain: &Chain) -> crate::Result<Unlock> {
+        Ok(Unlock::Signature(SignatureUnlock::new(Signature::Ed25519(
+            self.sign_ed25519(essence_hash, chain).await?,
+        ))))
+    }
 }
 
 /// An extension to [`SecretManager`].
@@ -258,22 +256,6 @@ impl SecretManage for SecretManager {
         }
     }
 
-    async fn signature_unlock(
-        &self,
-        input: &InputSigningData,
-        essence_hash: &[u8; 32],
-        metadata: &Option<RemainderData>,
-    ) -> crate::Result<Unlock> {
-        match self {
-            #[cfg(feature = "stronghold")]
-            Self::Stronghold(secret_manager) => secret_manager.signature_unlock(input, essence_hash, metadata).await,
-            #[cfg(feature = "ledger_nano")]
-            Self::LedgerNano(secret_manager) => secret_manager.signature_unlock(input, essence_hash, metadata).await,
-            Self::Mnemonic(secret_manager) => secret_manager.signature_unlock(input, essence_hash, metadata).await,
-            Self::Placeholder(secret_manager) => secret_manager.signature_unlock(input, essence_hash, metadata).await,
-        }
-    }
-
     async fn sign_ed25519(&self, msg: &[u8], chain: &Chain) -> crate::Result<Ed25519Signature> {
         match self {
             #[cfg(feature = "stronghold")]
@@ -365,9 +347,9 @@ impl SecretManager {
                         return Err(InputSelectionError::MissingInputWithEd25519Address)?;
                     }
 
-                    let block = self
-                        .signature_unlock(input, &hashed_essence, &prepared_transaction_data.remainder)
-                        .await?;
+                    let chain = input.chain.as_ref().ok_or(Error::MissingChain)?;
+
+                    let block = self.signature_unlock(&hashed_essence, chain).await?;
                     blocks.push(block);
 
                     // Add the ed25519 address to the block_indexes, so it gets referenced if further inputs have