+++ client/src/stronghold/secret.rs
@@ -13,8 +13,7 @@ use iota_stronghold::{
 };
 use iota_types::block::{
     address::{Address, Ed25519Address},
-    signature::{Ed25519Signature, Signature},
-    unlock::{SignatureUnlock, Unlock},
+    signature::Ed25519Signature,
 };
 use zeroize::Zeroize;
 
@@ -23,8 +22,7 @@ use super::{
     StrongholdAdapter,
 };
 use crate::{
-    api::RemainderData,
-    secret::{types::InputSigningData, GenerateAddressOptions, SecretManage},
+    secret::{GenerateAddressOptions, SecretManage},
     Error, Result,
 };
 
@@ -75,21 +73,6 @@ impl SecretManage for StrongholdAdapter {
         }
 
         Ok(addresses)
-    }
-
-    async fn signature_unlock(
-        &self,
-        input: &InputSigningData,
-        essence_hash: &[u8; 32],
-        _: &Option<RemainderData>,
-    ) -> Result<Unlock> {
-        let chain = input.chain.as_ref().unwrap();
-        let ed25519_sig = self.sign_ed25519(essence_hash, chain).await?;
-
-        // Convert the raw bytes into [Unlock].
-        let unlock = Unlock::Signature(SignatureUnlock::new(Signature::Ed25519(ed25519_sig)));
-
-        Ok(unlock)
     }
 
     async fn sign_ed25519(&self, msg: &[u8], chain: &Chain) -> Result<Ed25519Signature> {