+++ client/CHANGELOG.md
@@ -36,6 +36,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed `Client::get_output_ids_with_pagination()` to `Client::get_output_ids()`;
 - All MQTT related functions return an MQTT `Error`;
 - Re-export `mqtt` module instead of all its symbols;
+- `SecretManage::signature_unlock` is now auto-implemented in the trait;
+- `SecretManage::signature_unlock`'s parameter `&InputSigningData` replaced with `&Chain`;
+- `Message::SignatureUnlock`'s parameter `Box<InputSigningDataDto>` replaced with `Chain`;
+
+### Removed
+
+- `SecretManage::signature_unlock`'s `remainder` parameter;
+- `Message::SignatureUnlock`'s `remainder_data` field;
 
 ## 2.0.1-rc.7 - 2023-03-09